AutoMigrate handle all indexes , multiple and uniques, fkeys also handle on update with actions types

+++ core/orm/helpers.go
@@ -59,7 +59,7 @@ func linkModel[T comparable](to_table_name string, db *DatabaseEntity) {
 
 	go func() {
 		// rename field
-		handleRename(to_table_name, fields, cols, diff, db)
+		handleRename(to_table_name, fields, cols, diff, db,ftags)
 		wg.Done()
 	}()
 	wg.Wait()
@@ -84,16 +84,25 @@ func linkModel[T comparable](to_table_name string, db *DatabaseEntity) {
 func handleAddOrRemove[T comparable](to_table_name string, fields, cols, diff []string, db *DatabaseEntity, ftypes map[string]string, ftags map[string][]string) {
 	if len(cols) > len(fields) { // extra column db
 		for _, d := range diff {
+			if v,ok := ftags[d];ok && v[0] == "-"{
+				continue
+			}
 			fmt.Println(" ")
 			logger.Printfs("⚠️ found extra column '%s' in the database table '%s'", d, to_table_name)
+
 			statement := "ALTER TABLE " + to_table_name + " DROP COLUMN " + d
+
 			choice := input.Input(input.Yellow, "> do you want to remove '"+d+"' from database ? (Y/n): ")
 			if utils.SliceContains([]string{"yes", "Y", "y"}, choice) {
 				if len(databases) > 1 && db.Name == "" {
 					ddb := input.Input(input.Blue, "> There are more than one database connected, enter database name: ")
 					conn := GetConnection(ddb)
 					if conn != nil {
-						_, err := conn.Exec(statement)
+						_, err := conn.Exec("DROP INDEX IF EXISTS idx_"+to_table_name+"_"+d)
+						if logger.CheckError(err) {
+							return
+						}
+						_, err = conn.Exec(statement)
 						if err != nil {
 							temp := to_table_name+"_temp"
 							err := autoMigrate[T](db,temp)
@@ -112,7 +121,11 @@ func handleAddOrRemove[T comparable](to_table_name string, fields, cols, diff []
 				} else {
 					conn := db.Conn
 					if conn != nil {
-						_, err := conn.Exec(statement)
+						_, err := conn.Exec("DROP INDEX IF EXISTS idx_"+to_table_name+"_"+d)
+						if logger.CheckError(err) {
+							return
+						}
+						_, err = conn.Exec(statement)
 						if err != nil {
 							temp := to_table_name+"_temp"
 							err := autoMigrate[T](db,temp)
@@ -135,22 +148,43 @@ func handleAddOrRemove[T comparable](to_table_name string, fields, cols, diff []
 		}
 	} else if len(cols) < len(fields) { // missing column db
 		for _, d := range diff {
+			if v,ok := ftags[d]; ok && v[0] == "-"{
+				continue
+			}
 			fmt.Println(" ")
 			logger.Printfs("⚠️ column '%s' is missing from the database table '%s'", d, to_table_name)
 			choice, err := input.String(input.Yellow, "> do you want to add '"+d+"' to the database ? (Y/n):")
 			logger.CheckError(err)
 			statement := "ALTER TABLE " + to_table_name + " ADD " + d + " "
-			if ty, ok := ftypes[d]; ok {
+			if ty, ok := ftypes[d]; ok {			
+				res := map[string]string{}
+				fkeys := []string{}
+				indexes := []string{}
+				mindexes := map[string]string{}
+				uindexes := map[string]string{}
+				mi := &migrationInput{
+					dialect: db.Dialect,
+					fName: d,
+					fType: ty,
+					fTags: &ftags,
+					fKeys: &fkeys,
+					res: &res,
+					indexes: &indexes,
+					mindexes: &mindexes,
+					uindexes: &uindexes,
+				}
 				ty = strings.ToLower(ty)
 				switch {
 				case strings.Contains(ty, "str"):
-					res := map[string]string{}
-					fkeys := []string{}
-					handleMigrationString(db.Dialect, d, ty, &ftags, &fkeys, &res)
+					handleMigrationString(mi)
 					var s string
 					var fkey string
 					if v, ok := res[d]; ok {
 						s = v
+						if strings.Contains(v,"UNIQUE") {
+							s=strings.ReplaceAll(v,"UNIQUE","")
+							uindexes[d]=d
+						}
 					} else {
 						s = "VARCHAR(255)"
 					}
@@ -163,15 +197,17 @@ func handleAddOrRemove[T comparable](to_table_name string, fields, cols, diff []
 					}
 					statement += s
 				case strings.Contains(ty, "bool"):
-					res := map[string]string{}
-					fkeys := []string{}
-					handleMigrationBool(db.Dialect, d, ty, &ftags, &fkeys, &res)
+					handleMigrationBool(mi)
 					var s string
 					var fkey string
 					if v, ok := res[d]; ok {
 						s = v
 						if !strings.Contains(v,"DEFAULT 0") {
 							s += " DEFAULT 0"
+						}
+						if strings.Contains(v,"UNIQUE") {
+							s=strings.ReplaceAll(v,"UNIQUE","")
+							uindexes[d]=d
 						}
 					} else {
 						s = "INTEGER NOT NULL CHECK (" + d + " IN (0, 1)) DEFAULT 0"
@@ -185,13 +221,15 @@ func handleAddOrRemove[T comparable](to_table_name string, fields, cols, diff []
 					}
 					statement += s
 				case strings.Contains(ty, "int"):
-					res := map[string]string{}
-					fkeys := []string{}
-					handleMigrationInt(db.Dialect, d, ty, &ftags, &fkeys, &res)
+					handleMigrationInt(mi)
 					var s string
 					var fkey string
 					if v, ok := res[d]; ok {
 						s = v
+						if strings.Contains(v,"UNIQUE") {
+							s=strings.ReplaceAll(v,"UNIQUE","")
+							uindexes[d]=d
+						}
 					} else {
 						s = "INTEGER"
 					}
@@ -204,13 +242,15 @@ func handleAddOrRemove[T comparable](to_table_name string, fields, cols, diff []
 					}
 					statement += s
 				case strings.Contains(ty, "floa"):
-					res := map[string]string{}
-					fkeys := []string{}
-					handleMigrationFloat(db.Dialect, d, ty, &ftags, &fkeys, &res)
+					handleMigrationFloat(mi)
 					var s string
 					var fkey string
 					if v, ok := res[d]; ok {
 						s = v
+						if strings.Contains(v,"UNIQUE") {
+							s=strings.ReplaceAll(v,"UNIQUE","")
+							uindexes[d]=d
+						}
 					} else {
 						s = "DECIMAL(5,2)"
 					}
@@ -223,13 +263,15 @@ func handleAddOrRemove[T comparable](to_table_name string, fields, cols, diff []
 					}
 					statement += s
 				case strings.Contains(ty, "time"):
-					res := map[string]string{}
-					fkeys := []string{}
-					handleMigrationTime(db.Dialect, d, ty, &ftags, &fkeys, &res)
+					handleMigrationTime(mi)
 					var s string
 					var fkey string
 					if v, ok := res[d]; ok {
 						s = v
+						if strings.Contains(v,"UNIQUE") {
+							s=strings.ReplaceAll(v,"UNIQUE","")
+							uindexes[d]=d
+						}
 					} else {
 						if strings.Contains(db.Dialect, SQLITE) {
 							s = "TEXT"
@@ -258,6 +300,37 @@ func handleAddOrRemove[T comparable](to_table_name string, fields, cols, diff []
 					return
 				}
 
+				statIndexes := ""
+				if len(indexes) > 0 {
+					if len(indexes) > 1 {
+						logger.Error(mi.fName,"cannot have more than 1 index")
+					} else {
+						statIndexes = fmt.Sprintf("CREATE INDEX idx_%s_%s ON %s (%s)",to_table_name,d,to_table_name,indexes[0])
+					}
+				}
+				mstatIndexes := ""
+				if len(*mi.mindexes) > 0 {
+					if len(*mi.mindexes) > 1 {
+						logger.Error(mi.fName,"cannot have more than 1 multiple indexes")
+					} else {
+						if v,ok := (*mi.mindexes)[mi.fName];ok {
+							mstatIndexes = fmt.Sprintf("CREATE INDEX idx_%s_%s ON %s (%s)",to_table_name,d,to_table_name,d+","+v)
+						}
+					}
+				}
+				ustatIndexes := ""
+				if len(*mi.uindexes) > 0 {
+					if len(*mi.uindexes) > 1 {
+						logger.Error(mi.fName,"cannot have more than 1 multiple indexes")
+					} else {
+						if v,ok := (*mi.uindexes)[mi.fName];ok {
+							reste := ","
+							if v == "" {reste=v}
+							mstatIndexes = fmt.Sprintf("CREATE UNIQUE INDEX idx_%s_%s ON %s (%s)",to_table_name,d,to_table_name,d+reste+v)
+						}
+					}
+				}
+				
 				if utils.SliceContains([]string{"yes", "Y", "y"}, choice) {
 					if len(databases) > 1 && db.Name == "" {
 						ddb := input.Input(input.Blue, "> There are more than one database connected, database name:")
@@ -267,6 +340,33 @@ func handleAddOrRemove[T comparable](to_table_name string, fields, cols, diff []
 							if logger.CheckError(err) {
 								logger.Info(statement)
 								return
+							}
+							if statIndexes != "" {
+								_, err := conn.Exec(statIndexes)
+								if logger.CheckError(err) {
+									logger.Info(statIndexes)
+									return
+								}
+							}
+							if mstatIndexes != "" {
+								_, err := conn.Exec(mstatIndexes)
+								if logger.CheckError(err) {
+									logger.Info(mstatIndexes)
+									return
+								}
+							}
+							if ustatIndexes != "" {
+								_, err := conn.Exec(ustatIndexes)
+								if logger.CheckError(err) {
+									logger.Info(ustatIndexes)
+									return
+								}
+							}
+							if (Debug) {
+								logger.Printfs("ylstatement: %s",statement)
+								logger.Printfs("ylstatIndexes: %s",statIndexes)
+								logger.Printfs("ylmstatIndexes: %s",mstatIndexes)
+								logger.Printfs("ylustatIndexes: %s",ustatIndexes)
 							}
 							logger.Printfs("grDone, '%s' added to '%s', you may want to restart your server", d, to_table_name)
 						}
@@ -278,11 +378,38 @@ func handleAddOrRemove[T comparable](to_table_name string, fields, cols, diff []
 								logger.Info(statement)
 								return
 							}
+							if statIndexes != "" {
+								_, err := conn.Exec(statIndexes)
+								if logger.CheckError(err) {
+									logger.Info(statIndexes)
+									return
+								}
+							}
+							if mstatIndexes != "" {
+								_, err := conn.Exec(mstatIndexes)
+								if logger.CheckError(err) {
+									logger.Info(mstatIndexes)
+									return
+								}
+							}
+							if ustatIndexes != "" {
+								_, err := conn.Exec(ustatIndexes)
+								if logger.CheckError(err) {
+									logger.Info(ustatIndexes)
+									return
+								}
+							}
+							if (Debug) {
+								logger.Printfs("ylstatement: %s",statement)
+								logger.Printfs("ylstatIndexes: %s",statIndexes)
+								logger.Printfs("ylmstatIndexes: %s",mstatIndexes)
+								logger.Printfs("ylmstatIndexes: %s",mstatIndexes)
+							}
 							logger.Printfs("grDone, '%s' added to '%s', you may want to restart your server", d, to_table_name)
 						}
 					}
 				} else {
-					logger.Info("Nothing changed")
+					logger.Printfs("grNothing changed")
 				}
 			} else {
 				logger.Info("case not handled:", ty, ftypes[d])
@@ -292,12 +419,15 @@ func handleAddOrRemove[T comparable](to_table_name string, fields, cols, diff []
 }
 
 // handleRename handle sync with db when renaming fields struct
-func handleRename(to_table_name string, fields, cols, diff []string, db *DatabaseEntity) {
+func handleRename(to_table_name string, fields, cols, diff []string, db *DatabaseEntity,ftags map[string][]string) {
 	// rename field
 	old := []string{}
 	new := []string{}
 	if len(fields) == len(cols) && len(diff)%2 == 0 && len(diff) > 0 {
 		for _, d := range diff {
+			if v,ok := ftags[d];ok && v[0] == "-"{
+				continue
+			}
 			if !utils.SliceContains(cols, d) { // d is new
 				new = append(new, d)
 			} else { // d is old
